requirements.txt
litellm
pre-commit
datasets
tiktoken
libcst
llama-index
jsonlines
tomli
python-dotenv
swebench @ git+https://github.com/princeton-nlp/SWE-bench
